#3941 `shake`, `sodg` tests that use `Farea` to integrations
